+++ jinny-l/week_20/BOJ_15649_N과_M_1.java
@@ -1,4 +1,4 @@
-package week_19;
+package week_20;
 
 import java.io.BufferedReader;
 import java.io.IOException;
